requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# File Handling
Pillow==10.1.0

# PDF Generation
WeasyPrint==60.1
reportlab==4.0.7

# API Documentation
drf-spectacular==0.26.5

# Filters
django-filter==23.3

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
